backend/requirements.txt
# Core
torch>=2.0.0
torchvision>=0.15.0
timm>=1.0.0
Pillow>=9.0.0

# ML Infrastructure
modal>=1.0.0

# APIs
requests>=2.28.0
python-dotenv>=1.0.0

# Web Server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
motor>=3.3.0
pymongo>=4.5.0
sentence-transformers>=2.2.0

# HTTP Client
httpx>=0.26.0

# Utilities
tqdm>=4.65.0
numpy>=1.24.0
